using a dirty snapshot
Making use of RelidByRelfilenodeDirty(), a
rewritten version of RelidByRelfilenode that is using a dirty snapshot.

That way we avoid false positive for files that have been created, associated transactions
still in progress and their relfilenode still not visible in pg_class.

Before this commit we were checking for locks instead (see with_lock_check branch).