don't report regular relation first segment with size zero as orphaned.
After a drop relation the first segment is truncated and kept (all the other segments are unlinked).
This is an intentional behavior see [0].

It's up to the next checkpoint to unlink it.
Even if this file is really orphaned (means with no related entry in pg_class) and that the checkpointer would not complain
if we remove it, we don't want to report it as being orphaned.

The reason behind is that if we remove it (because reported as orphaned), that would break the protection it is made for, see [0].

So don't report it, and let the next checkpoint remove it instead.

[0]: https://github.com/postgres/postgres/blob/REL_12_8/src/backend/storage/smgr/md.c#L225