Compare with last checkpoint time for empty first segment
In 5e452efe47a32647585eb997729ad4497a4d6a90 we don't report regular relation
first segment with size zero as orphaned (see the commit message for explanation).

That does make sense if the file modification time is after the last checkpoint
but not that much if it is older.

This commit reports regular relation first segment with size zero as orphaned if
it is orphaned and its modification time is older than the last checkpoint.